factor out a common port settings saga function missed in #4528
